fix(tests): EIP-7702 test_ext_code_on_set_code
